Take a TcpStream by argument rather than connecting internally
*Note: Info command currently contains a hack as it turns out to rely on
the `read_exact` call ending when the firmware closes the socket. It
does this as it does not know the exact length of the incoming message.*

This gives the user some freedom in terms of how the approach used to
constructing the stream, timeouts, etc.

Also returns an error when failing to write to the stream, rather than
panicking internally.